requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
sqlalchemy>=2.0
asyncpg
alembic  # for migrations

# Observability
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-sqlalchemy
opentelemetry-exporter-prometheus
opentelemetry-instrumentation-logging
